Deadline point/staging version

+++ app/src/main/java/com/example/cooptesapp/models/db/BarcodeEntity.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "barcode")
 data class BarcodeEntity(
     @PrimaryKey(autoGenerate = true)
-    val id: Int = 0,
-    val pack_id: Int,
+    val id: Long = 0,
+    val pack_id: Long,
     val body: String
 )
